+++ internal/handlers/handler_oauth2_authorization_consent_explicit.go
@@ -62,6 +62,12 @@ func handleOAuth2AuthorizationConsentModeExplicitWithID(ctx *middlewares.Autheli
 		return nil, true
 	}
 
+	if !consent.Subject.Valid && consent.Subject.UUID == uuid.Nil {
+		handleOAuth2AuthorizationConsentRedirect(ctx, issuer, consent, client, userSession, rw, r, requester)
+
+		return nil, true
+	}
+
 	if subject != consent.Subject.UUID {
 		ctx.Logger.Errorf(logFmtErrConsentSessionSubjectNotAuthorized, requester.GetID(), client.GetID(), client.GetConsentPolicy(), consent.ChallengeID, userSession.Username, subject, consent.Subject.UUID)
