+++ web/src/views/LoginPortal/AuthenticatedView/AuthenticatedView.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 
-import { Button, Theme } from "@mui/material";
+import { Theme } from "@mui/material";
 import Grid from "@mui/material/Grid";
 import { useTranslation } from "react-i18next";
 import { makeStyles } from "tss-react/mui";
 
-import { LogoutRoute as SignOutRoute } from "@constants/Routes";
-import { useRouterNavigate } from "@hooks/RouterNavigate";
+import LogoutButton from "@components/LogoutButton.js";
 import MinimalLayout from "@layouts/MinimalLayout";
 import { UserInfo } from "@models/UserInfo";
 import Authenticated from "@views/LoginPortal/Authenticated";
@@ -18,13 +17,7 @@ export interface Props {
 const AuthenticatedView = function (props: Props) {
     const { t: translate } = useTranslation();
 
-    const navigate = useRouterNavigate();
-
     const { classes } = useStyles();
-
-    const handleLogoutClick = () => {
-        navigate(SignOutRoute);
-    };
 
     return (
         <MinimalLayout
@@ -34,9 +27,7 @@ const AuthenticatedView = function (props: Props) {
         >
             <Grid container direction={"column"} justifyContent={"center"} alignItems={"center"}>
                 <Grid size={{ xs: 12 }}>
-                    <Button id={"logout-button"} color={"secondary"} onClick={handleLogoutClick}>
-                        {translate("Logout")}
-                    </Button>
+                    <LogoutButton />
                 </Grid>
                 <Grid size={{ xs: 12 }} className={classes.mainContainer}>
                     <Authenticated />
