+++ web/src/views/LoginPortal/SecondFactor/SecondFactorForm.tsx
@@ -7,6 +7,7 @@ import { useTranslation } from "react-i18next";
 import { Route, Routes } from "react-router-dom";
 import { makeStyles } from "tss-react/mui";
 
+import LogoutButton from "@components/LogoutButton";
 import {
     SecondFactorPasswordSubRoute,
     SecondFactorPushSubRoute,
@@ -14,7 +15,6 @@ import {
     SecondFactorWebAuthnSubRoute,
     SettingsRoute,
     SettingsTwoFactorAuthenticationSubRoute,
-    LogoutRoute as SignOutRoute,
 } from "@constants/Routes";
 import { useLocalStorageMethodContext } from "@contexts/LocalStorageMethodContext";
 import { useNotifications } from "@hooks/NotificationsContext";
@@ -81,10 +81,6 @@ const SecondFactorForm = function (props: Props) {
         }
     };
 
-    const handleLogoutClick = () => {
-        navigate(SignOutRoute);
-    };
-
     const showMethods = props.factorKnowledge && props.configuration.available_methods.size > 1;
 
     return (
@@ -104,9 +100,7 @@ const SecondFactorForm = function (props: Props) {
             ) : null}
             <Grid container direction={"column"} justifyContent={"center"} alignItems={"center"}>
                 <Grid size={{ xs: 12 }}>
-                    <Button id={"logout-button"} color={"secondary"} onClick={handleLogoutClick}>
-                        {translate("Logout")}
-                    </Button>
+                    <LogoutButton />
                     {showMethods ? " | " : null}
                     {showMethods ? (
                         <Button id={"methods-button"} color="secondary" onClick={handleMethodSelectionClick}>
